backendRiff/backendRiffRoot/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
audioread==3.0.1
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
decorator==5.1.1
distlib==0.3.8
dj-rest-auth==7.0.1
Django==4.2.16
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.10
joblib==1.4.2
jwcrypto==1.5.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.2
pipenv==2023.11.17
platformdirs==4.1.0
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
PyJWT==2.9.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
soundfile==0.13.1
soxr==0.5.0.post1
spotipy==2.25.0
sqlparse==0.5.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.25.0
